Added a draft README.md. Cleaned up the code and deleted sections related to compute shader experiments. Made some minor code fixes.

+++ undoManager.cpp
@@ -44,7 +44,7 @@ void undoManager::undo(int actionsToUndo)
 	// Take original data.
 	std::vector<VertexData> copyOfOriginalData = currentPointCloud->originalData;
 	
-	for (size_t i = undoActions.size() - 1; i >= 0; i--)
+	for (int i = undoActions.size() - 1; i >= 0; i--)
 	{
 		if (undoActions[i]->affectedPointCloud != currentPointCloud)
 			continue;
